Update grid code copyable textarea (HTML + CSS)
Previously, the textarea containing the grid HTML contained HTML with inline styling. When tested, it was a mess to copy+paste due to the sheer volume of text (When I tried to paste all that HTML with inline styling into a new project in VSCode, it's format-on-save feature and the Prettier extension couldn't handle the amount of text, and either took up to a minute to process or just straight-up crashed.)

This update removes the inline styling from the HTML and instead adds another textarea for styling with CSS - so now the user can copy both the provided HTML and CSS into their project and get the same results as before, but in a much better way.