Update manner of hiding/unhiding error-msg
Previously I used display:none;/display:block; to hide the error message. This commit changes it to visibility:hidden;/visibility:visible because I decided I want to maintain it's spacing.

Previously, when the error message was shown the section would expand vertically which is a bit jarring. Using the the visibility property ensures the section doesn't expand when the error message is shown.